chore :: 파일구조 분해 및 Stack 설정
