+++ package.json
@@ -9,11 +9,17 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/stack": "^6.4.0",
     "expo": "~51.0.14",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
     "react-native": "0.74.2",
-    "react-native-safe-area-context": "^4.10.5",
+    "react-native-gesture-handler": "^2.17.1",
+    "react-native-reanimated": "^3.13.0",
+    "react-native-safe-area-context": "^4.10.8",
+    "react-native-screens": "^3.32.0",
     "react-native-webview": "^13.10.4"
   },
   "devDependencies": {
